@@ -14,17 +14,21 @@
   "author": "Danude Sandstorm",
   "license": "MIT",
   "dependencies": {
-    "@loadable/component": "^5.10.3",
-    "@types/react": "^16.9.9",
-    "@types/react-dom": "^16.9.2",
-    "lokijs": "^1.5.7",
-    "mobx": "^5.13.1",
+    "@loadable/component": "^5.11.0",
+    "@material-ui/core": "^4.5.1",
+    "@material-ui/icons": "^4.5.1",
+    "@material-ui/styles": "^4.5.0",
+    "@types/react": "^16.9.15",
+    "@types/react-dom": "^16.9.4",
+    "@types/react-router-dom": "^5.1.2",
+    "lokijs": "^1.5.8",
+    "mobx": "^5.15.0",
     "mobx-react": "^5.4.4",
     "prop-types": "^15.7.2",
-    "react": "^16.10.1",
-    "react-collapsible": "^2.6.0",
+    "react": "^16.12.0",
+    "react-collapsible": "^2.6.2",
     "react-digit-input": "^1.0.0",
-    "react-dom": "^16.10.1",
+    "react-dom": "^16.12.0",
     "react-interactive": "^0.9.1",
     "react-loadable": "^5.5.0",
     "react-onclickoutside": "^6.9.0",
@@ -36,37 +40,51 @@
     "whatwg-fetch": "^3.0.0"
   },
   "devDependencies": {
-    "@babel/cli": "^7.6.2",
-    "@babel/core": "^7.6.2",
-    "@babel/node": "^7.6.2",
-    "@babel/plugin-proposal-class-properties": "^7.5.5",
-    "@babel/plugin-proposal-decorators": "^7.6.0",
-    "@babel/plugin-proposal-object-rest-spread": "^7.6.2",
-    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
-    "@babel/plugin-transform-runtime": "^7.6.2",
-    "@babel/polyfill": "^7.6.0",
-    "@babel/preset-env": "^7.6.2",
-    "@babel/preset-flow": "^7.0.0",
-    "@babel/preset-react": "^7.0.0",
-    "@babel/preset-typescript": "^7.6.0",
-    "@babel/register": "^7.6.2",
-    "@babel/runtime": "^7.6.2",
+    "@babel/cli": "^7.7.5",
+    "@babel/core": "^7.7.5",
+    "@babel/node": "^7.7.4",
+    "@babel/plugin-proposal-class-properties": "^7.7.4",
+    "@babel/plugin-proposal-decorators": "^7.7.4",
+    "@babel/plugin-proposal-dynamic-import": "^7.7.0",
+    "@babel/plugin-proposal-object-rest-spread": "^7.7.4",
+    "@babel/plugin-transform-computed-properties": "^7.7.4",
+    "@babel/plugin-transform-runtime": "^7.7.5",
+    "@babel/polyfill": "^7.7.0",
+    "@babel/preset-env": "^7.7.5",
+    "@babel/preset-flow": "^7.7.4",
+    "@babel/preset-react": "^7.7.4",
+    "@babel/preset-typescript": "^7.7.4",
+    "@babel/register": "^7.7.4",
+    "@babel/runtime": "^7.7.1",
     "babel-eslint": "^10.0.3",
     "babel-loader": "^8.0.5",
-    "css-loader": "^3.2.0",
-    "eslint": "^6.5.1",
+    "css-loader": "^3.2.1",
+    "eslint": "^6.7.2",
+    "eslint-plugin-flowtype": "^4.3.0",
     "eslint-plugin-import": "^2.18.2",
-    "eslint-plugin-react": "^7.15.0",
-    "eslint-plugin-react-hooks": "^2.1.1",
+    "eslint-plugin-react": "^7.17.0",
+    "eslint-plugin-react-hooks": "^2.3.0",
     "mini-css-extract-plugin": "^0.8.0",
-    "node-sass": "^4.12.0",
+    "node-sass": "^4.13.0",
     "optimize-css-assets-webpack-plugin": "^5.0.3",
     "sass-loader": "^8.0.0",
-    "style-loader": "^1.0.0",
-    "typescript": "^3.6.4",
+    "style-loader": "^1.0.1",
+    "typescript": "^3.7.3",
     "uglifyjs-webpack-plugin": "^2.2.0",
-    "webpack": "^4.41.0",
-    "webpack-cli": "^3.3.9",
-    "webpack-dev-server": "^3.8.1"
+    "webpack": "^4.41.2",
+    "webpack-cli": "^3.3.10",
+    "webpack-dev-server": "^3.9.0"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
